requirements.txt
# selenium
# python-dotenv
# langchain
# langchain_community
# langchain-cli
# langserve
# langchain-huggingface
# transformers
# bitsandbytes
# fastapi
# huggingface_hub
# beautifulsoup4
# accelerate
# qdrant-client

python-dotenv
langchain
langchain_community
langserve
fastapi
beautifulsoup4
qdrant-client
google-generativeai
langchain-google-genai
